Changed create logic, removed global families and unified register transmission protocol.
Global family IDs are no longer used. They used be used for arbitration among the family tables of the various processor. In the new method, a create 
is forwarded only after the local entry has been allocated. Thus the previous FID is known to the next core, and a previous and next link can be 
maintained on each core for local communication. This has as added benefits that the create becomes faster on the network, since reservation and 
unreservation are no longer necessary.

Transmitting globals has also been removed from the create process. Because multiple threads can now wait on a single register, globals are now 
requested when they are used, just like shareds. Without the processing of globals in the create, the create logic becomes much simpler as well.
To allow shareds (and, incidentally, globals) to be written after the create, remote waiting information had to be stored in the register, so writes 
to that register can trigger a transmission of the shared or global. This should also allow memory loads to be done on (child) shareds.

The register transmission protocol has been simplified and clarified to support this, as well as prepare for sending and receiving globals and 
shareds.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@1928 e97e5017-a994-416e-809d-76780e9f78db