mgsim: fixed the floating point shareds bug.
The bug was a deadlock caused by floating point operations clearing the destination register, thus clearing the remote request information for the shared. This was fixed by not clearing the remote request information when clearing a register.

Then the issue arose that a long latency operation to a parent shared would not send the shared in the end, because the FPU and D-Cache did not have the information from the pipeline, that the register would have to be forwarded. This was solved by storing the remote request information in the EMPTY writeback in the pipeline when the long latency operation was started. This, in effect, pretends that a remote request has been made on the register. So the writeback then forwards as usual.

Then there were also some port permissions and small logic bugs to fix with these new additions.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2104 e97e5017-a994-416e-809d-76780e9f78db