mgsim: implemented a static configuration area for resource management.
After loading all code and data, the simulator will put the place information (from NumProcessors) into an area of memory and put a pointer to that memory in the Local #2 register to the main thread.
The memory format of the region is:

uint32_t count
uint32_t cores[*]

Each 'cores' element describes a core as: (place << 16) | (pid << 0). The low 16 bits are the core ID for use in PlaceIDs for delegated creates. The high 16 bits can be used to know which cores belong to which place. Cores in the same place havv the same 'place' number.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2111 e97e5017-a994-416e-809d-76780e9f78db