mgsim: fixed exclusive self delegation bug
Performing an exclusive delegated create to the same core would cause a deadlock because the delegated create is treated as an exclusive create twice; once at the issued create and once at the 
received delegated create. This situation was not anticipated. However, delegating to a fixed core that may be the same is a possible scenario. Therefore, the assertions concerning this scenario have 
been removed and locally issued exclusive delegated creates are not stored on the exclusive create queue, thus avoiding the double exclusive create.

A regression test case has also been added.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2178 e97e5017-a994-416e-809d-76780e9f78db