mgsim: fixed infinite pipeline waiting loop regression test case
The test case used the "_start" label for the main program. However, the linker during 'make check' assumes the entry label is "main". Since this label is missing, the entry point is assumed 0, instead of 4, what it should be. This causes a wrong value for GP to be calculated, causing a read outside of valid memory.

The entry point label has been changed to main



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2179 e97e5017-a994-416e-809d-76780e9f78db