mgsim: banked memories now use NumProcessors for default number of banks
With the previous configuration system change, the BankedMemories used a fixed configuration value as default for the number of banks. This has been changed to use the total of the NumProcessors list as default. This way, when the MemoryBanks option is left out, it default to the number of cores in the system.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2194 e97e5017-a994-416e-809d-76780e9f78db