mgsim: implemented variable FP operation-to-pipeline mapping
The new system allows a pipeline to be shared between operations or to have multiple pipelines for the same operation. This change was effected because for Add and Sub it makes more sense to have two shared Add/Sub pipelines rather than one Add and one Sub pipeline, since both operations are so similar.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2202 e97e5017-a994-416e-809d-76780e9f78db