mgsim: fixed bug with memory writeback to shared
Due to a missing entry in the whitelist for the port on the network register that stores outgoing shareds, a writeback from the D-Cache to a register that is a shared that is being waited on, triggered an assertion (or deadlock). The correct entries have been added to the whitelist.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2269 e97e5017-a994-416e-809d-76780e9f78db