mgsim: restructured and improved debugging code
The debug code was previously all contained in commands.cpp/h and required public accessors for all important data fields on all objects. This has been restructured so that every component now has its own command handler which can access its own private data. This change better modularizes components (most code related to a component is now just in its class).

The debug handlers were also improved by adding range parameters to the large tables (threads, families, registers), adding a data dump to the caches, and improving layout and information for other components. A debug handler for the FPU has also been added.

The RandomBanked memory has also been changed to simply inherit from the Banked memory and override the address calculation, instead of having lots of code duplication.


Exceptions from the simulator have also been improved by adding details about the component that threw the exception, and which component was executing when the exception was thrown. When an exception occurs through the pipeline, the originating thread, family and instruction PC are listed as well.

Deadlock reports have been extended with a small summary of the state of the system.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2314 e97e5017-a994-416e-809d-76780e9f78db