mgsim: added LDFP instruction support
Because we do not want to write the index to L0 AND the frame pointer to L1 during thread creation, only L0 is written and the frame pointer can be obtained by the program with the LDFP 
instruction, which takes a single destination register.

Also, support for SETREGS was removed. It was kept for legacy compatibility until all software switched to the new mechanism, which is providing the same values at the ALLOCATE 
instruction.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2462 e97e5017-a994-416e-809d-76780e9f78db