mgsim: changed the core design of the simulator and various other tweaks
Previously, on every cycle, the simulator would call all components who would then check if they had work to do. This typically involved checking if a buffer was full and if not, just return a value indicating nothing could be done. To avoid writes being picked up by reads in the same cycle, thus lending to the possibility that events could propagate unrealistically fast through a system depending on the order in which components are called, each cycle was split into two phases, a Read phase and a Write phase, which meant components would be essentially called twice.

This mechanism was replaced with a run-queue system. Each _process_ in a component is sensitive on one or more _storage_ objects, like buffers or linked lists. When something writes to these storage objects, it adds the process that is sensitive on it to the run-queue. This ensures that only those processes that have work to do, are called. The read/write cycle split was also removed by buffering writes to storage objects until the end of the cycle. At that time, all storages that have been written to get updated so the new value is available to be read in the next cycle.

A similar mechanism has been applied to all arbitrators; an arbitrator is only considered for arbitration when a request has been made to it. And, since all these mechanisms use linked lists, there is only a O(1) overhead for registering arbitrations, updates and new processes.


Because of this change in the core design, some things in the simulator had to be slightly redesigned or changed:

The pipeline and FPU are now a single stage each. It used to be split up in phases and under the old system it would guarantee the order in which the processes in a component were called, so the pipeline could be handled back-to-front. Under the new system, this wasn't possible, so the entire pipeline has been merged into a single process.

The new storage objects exposed some conditions where there were multiple writes to the same object without arbitration, so arbitration had to be added. Along with this, some buffers have been added in the D-Cache and I-Cache to simplify the arbitration sources and, as a side effect, make the memory interface more realistic. Overall, the addition of buffers and arbitrators have been the simulation overall more realistic (although a bit slower).

Also as a result of this, L0 and L1 can no longer be written in the same cycle, so the writing of L1 has been removed. Note that the same information can be obtained via the recently introduced LDFP instruction.

The RegisterFile, now being an entire storage object (although not a sensitive storage object), buffers all writes for the update phase. Since the buffer size is the same as the number of ports (maximum only one write per port per cycle), so this provides another layer of sanity checking.

Arbitrators now also sanity check that processes don't request access more than once in a cycle, which would indicate different parts of the code wanting to use the same resource.


The pipeline's Read Stage has been simplified and made more robust. Previously, for multi-register operands, it would try to get as many parts of the operand from the bypasses as it could, possibly filling multiple different parts in a single cycle. Keeping track of the proper instruction dependencies and instruction replay was non-trivial. This has been changed such that the Read stage now reads the parts of an operand one part at a time, in order. So, even if different parts of the operand are available from the bypass bus, it only looks for its current operand part. This greatly simplifies the code since only one register is ever considered at a cycle and makes the whole thing generally more robust, understandable and maintainable.

Deadlock detection was also improved a bit by noting which component is reporting which messages.

As a final note, the profiling code has been removed because it was only used in the code to allow the main kernel loop to be timed, but this has rarely to never been used and for proper profiling gprof can be used anyway. Along with the profiling code, the 'simtime' code has been removed as well.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2465 e97e5017-a994-416e-809d-76780e9f78db