mgsim: fixed several bugs and made some tweaks
The first bug that was fixed was in the Network code. Like the fix in [2477], this bug caused the simulator to hang infinitely when tokens were being 
requested, but the create process has stalled. This is essentially deadlock, but the Network code was written such that, as long as the token was 
requested, it would return SUCCESS. This has been changed to return FAILED in the appropriate cases.

The second bug had to do with the propagation of parent shareds and the synchronization token. In certain (not too uncommon) situations, the latter 
could overtake the former on the ring network, causing families to be cleaned up before their entries could be used to forward the parent shared 
(because each family entry contains the FID of the corresponding family on the next, and previous, core). Several solutions were possible:

* Send synch tokens and parent shareds on the same network, guaranteeing the correct order. This would, however, increase pressure on the channel and 
block unrelated syncs from proceeding in the face of many shareds.

* Count off the parent shareds on every core between the core running the last thread and the parent core. Unfortunately, this counter would have to 
be decremented after actually sending the shared from the outgoing buffer, where the information about the FID on the current core would be lost.

* The implemented solution came after the realization that we have a field in the Family Table for the FID on the parent core anyway. If this field is 
always set properly, the parent shared could be sent with that FID in the first place, no longer requiring the intermediate family entries, still 
leaving the synch channel and the register channel separate.

The latter solution also required adding a LPID field to the register address information, to know what core the parent was located on, once the 
parent shared was in transit (that information is stored in the family table, but that could no longer be used). The 'incoming register response' 
processes for group and remote network were split since they both required different behavior. This was also due to the fact that register propagation 
was rewritten so global and shared requests for delegated families now go entirely to the local parent core first, which then forwards them onto the 
remote network. And vica versa for their replies.

Lastly, some small changes were also made some debug output and the DecodeStage. It was also confusing that the family contained a 'shareds' and 
'globals' pointer which would be expected to point to the parent shareds, or be invalid otherwise. However, they were set to the local caches if the 
family wasn't running on the parent core. Changing this to the expected behavior also simplified some other code here and there.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2513 e97e5017-a994-416e-809d-76780e9f78db