mgsim: added dummy pipeline stage
The pipeline can now be extended with a variable number of dummy stages. These stages sit between the Memory and Writeback stages and just forward their inputs. The purpose of these stages is to be able to test performance for different pipeline lengths, as well as test correctness and robustness of the implementation for differing delays between Read/Execute and Writeback.

The parameter that defines the number of these stages in the pipeline is called "NumPipelineDummyStages" and is by default set to 0 to mimic the old behavior.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2523 e97e5017-a994-416e-809d-76780e9f78db