mgsim: implemented divide instructions
Since the Alpha lacks divide operations, divisions and modulos must be performed with a function call. Since this produces some headaches in our research, it was decided to hack in divide instructions and get rid of these function calls.

Four new instructions have been added to the simulator, for long/quad and signed/unsigned division. Modulo can be performed in the program by using div, mul and sub.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2668 e97e5017-a994-416e-809d-76780e9f78db