mgsim: allowed writes to loading registers
As described in #198, we want to allow writes to registers which have been the target of a memory load. Since we cannot overwrite the register, we could stall. However, this ties up the pipeline port, and because it has exclusive priority, the D-Cache can never write back, thus causing deadlock.

The adopted solution is to stall, but only try to write again every other cycle, thus giving the D-Cache an opportunity to write the results.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2705 e97e5017-a994-416e-809d-76780e9f78db