mgsim: moved Place ID from SETPLACE to ALLOCATE
The place ID for a create has been moved from the SETPLACE instruction to the ALLOCATE instruction. The place ID must now be set in the register that allocate takes. This used to be an output-only register but is now an input/output register. The SETPLACE instruction is now obsolete.

This was done so that the place information will be available at the allocation of a context, which allows the hardware to do more intelligent things.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2728 e97e5017-a994-416e-809d-76780e9f78db