mgsim: added runtime profiling
Memory loads from the top 256 bytes (minus the top 32, which implement SETEMPTY) now return runtime system-wide profiling information. The profiling block is an array of integers (integer size depends on architecture) and its elements currently represent:

0: Number of instructions executed
1: Number of floating-point instructions executed

The rest of the array is reserved for future use and should not be used.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2833 e97e5017-a994-416e-809d-76780e9f78db