mgsim: fixed D-Cache coherency bug
D-Cache coherency was flawed in the previous version. A write on a core would not affect loads in progress on other cores, causing old data to be cached on those cores. This has been fixed by adding a bitmap to each cache-line, snooping the bus and writing data into loading lines with mask, and merging read data with based on the mask.

Some of cache-line code has also been simplified; the PROCESSING state has been merged into the LOADING line since there's no point to differentiate the two states.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2884 e97e5017-a994-416e-809d-76780e9f78db