mgsim: fixed register context depletion bug
On group creates, reserved register contexts are released when the registers are released. This presents a problem when no registers of a certain type are required. The registers are reserved, but never allocated and freed. Thus, the core would run out of free register contexts.

This was fixed by releasing reserved register contexts if the core didn't require them (i.e., the allocation size is 0).


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2888 e97e5017-a994-416e-809d-76780e9f78db