mgsim: fixed function call bug (Alpha)
Plain old function calls on the Alpha use the BSR instruction which stores the return PC in a specified register. On the microgrid, if the next instruction is the control word, the address of the instruction after that one is stored instead. This test, however, was wrong, causing returns from BR/BSR to frequently skip the instruction after it.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2915 e97e5017-a994-416e-809d-76780e9f78db