mgsim: moved runtime profiling address range
The runtime profiling address range was previously at the top of the address space. Unfortunately, this happens to be in the TLS area, causing problems for the last thread on the last processor in a system (since it's TLS contains this region).

Therefore, the profiling range has been moved to the bottom of the address space so it's in the Global Space, and can still be efficiently used. Addresses between 4 and 256 are now 'special': loads from address 4 - 8 are the new setempty instruction and loads from 8 - 256 are the profiling range. 0 - 4 has been left as normal to keep NULL pointers invalid.

The ELF loader code has been adjusted to verify that no ELF sections overlap with reserved regions (TLS, profiling range). Support for flat binaries loaded at address 0 has been removed.

Because of this change, the example programs need to be linked at an address larger than the first page. The linker script and README example have been modified accordingly.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@2961 e97e5017-a994-416e-809d-76780e9f78db