mgsim: cleaned up graphics code
The graphics code is now merged and simplified into display.cpp/h. The system now defines the display and passes it to each processor to pass it on to the execute stage which can do the resize and putpixel operations. The kernel gets a reference to the display in order to check for events every cycle.

The refresh has been reimplemented with a timer which means it now has a real-time refresh rate instead of a simulated cycle-based refresh rate.

The display width and height have been removed from the configuration, as the program determines the frame buffer size and the configuration just specifies the scale factors.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3111 e97e5017-a994-416e-809d-76780e9f78db