mgsim: fixed a few bugs and implemented FP optimizations
A fix in r3128 was made unjustly, the original code was correct and has been restored.

The optimizations in r3149 were incorrect in two methods and have been fixed.

The FP optimization is not to use the IEEE-754 emulation class unless the --enable-softfpu option is passed to configure. Instead, the native types of float and double are used in a union with an integer to obtain the IEEE-754 binary representation of floats. This assumes that the host architecture implements floats and doubles according to IEEE-754, which is usually the case.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3152 e97e5017-a994-416e-809d-76780e9f78db