[mgsim] fixed livelock bug in COMA
A bug in the COMA protocol caused an eviction and a token kill request to chase around the ring, never actually being evicted.This was caused because the protocol sends out a token kill request even if the cache-line with all tokens is being moved.In such a case, the cache-line has more than the maximum number of tokens and, when evicted before the token kill request comes full circle, will cause the livelock scenario. The solution was to not send out a token kill request if the evicted cache-line has all the tokens.

Some improvements to tracing messages have also been made parallel to fixing this bug.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3538 e97e5017-a994-416e-809d-76780e9f78db