[mgsim] added arbitration to ParallelMemory
Although the ParallelMemory has a dedicated input buffer for every processor, each processor still has two processes (ICache and DCache) that can simultaneously write to the processor's buffer. Since there was no arbitration on this buffer, the simulator would assert because of too many pushes in a single cycle. This has been fixed by adding an arbitrator to the port's buffer.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3624 e97e5017-a994-416e-809d-76780e9f78db