[mgsim] Fixed branch optimisation
The branch optimisation in [3629] was flawed in that the next PC is only written if the target is not the next instruction. This defeats the purpose of using the branch to get the current PC.

The fix involves that the register is always written (if the branch is taken), and the PC is only updated and the pipeline is only flushed if the target is not the next instruction.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3632 e97e5017-a994-416e-809d-76780e9f78db