[mgsim] Fixed branch/thread end bug
A branch instruction annotated with a thread end should ignore the end if the branch is taken as the end logically follows the branch and is thus never reached.
However, with the changes in [3692], [3632] and [3633], the kill was only ignored along with the forced switch, i.e. if the branch is not to the next PC.
Since we want to ignore the kill if the branch is taken regardless of the destination PC, the clearing of the kill flag has been moved out of the optimization check.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3683 e97e5017-a994-416e-809d-76780e9f78db