[mgsim] Removed broken optimization for delegation creates (fixes #263)
Delegation allocations to the same core are optimized into a local create. Unfortunately, this meant that they both ran locally, and has a PLACE_LOCAL property, which meant that default subfamilies inherited that place, and also ran locally.

The desired behavior is for delegated creates to run locally, but have GROUP_PLACE (so their children are group creates). This behavior is correct in delegated creates to other cores.

To fix the exceptional case, the optimization that turned delegations to the same core into local creates has been removed. This has the side effect that such delegations now (needlessly) go through the network.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3685 e97e5017-a994-416e-809d-76780e9f78db