[mgsim] Fixed bug with jumps in Alpha ISA
The jump and branch instructions in the Alpha ISA store the return address in a specified register. However, to ensure correctness of calculating GP, if this address lies at the beginning of a cacheline (so the instruction was the last in a cacheline), the control word needs to be skipped in the address calculation. This was only done for branches, and not for jumps, meaning in rare cases GP was calculated wrong. This has been fixed.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3769 e97e5017-a994-416e-809d-76780e9f78db