[mgsim] Fixed group creates under heavy register usage
Group creates when the register usage is high could result in a worse-than-should-be performance. This was caused by the register allocation on the creating core occuring before the distribution of the create. If this register allocation uses the last registers on the core, the group distribution fails because it thinks there are not enough registers on the ring to run the create. The family thus gets reinitialized to a local family.

To fix this, register allocation has been moved to after the create distribution in the create process.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3789 e97e5017-a994-416e-809d-76780e9f78db