[mgsim] Fixed thread exhaustion bug
Group creates were bugged and resulted in an additional thread being reserved. After a while, this would make the entire thread table reserved, causing family allocations to fail due to lack of free contexts. The fix for this fixes #267.

Besides this fix, the thread and family table and register allocation unit have been extended with more sanity checking and printing of the context counts.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3815 e97e5017-a994-416e-809d-76780e9f78db