[mgsim] Added COMA L2 cache hit/miss statistics
Just to clarify, we interested in hits and misses for reads only.

A hit for a read is when the data is in the cache. When the line is empty or being loaded, it's a miss. This definition follows from the 'intuitive' implication that a hit means that the result is returned immediately, whereas on a miss you have to wait.

With a write you don't have to wait for anything. Writes are essentially fire-and-forget, except we need the consistency confirmation (which you always have to wait for, unless you have the only copy, but then it's more a measure of exclusivity). So in my opinion, to incorporate writes into hit/miss statistics would only obscure the results that we care about -- those for reads.

The L1 DCache also only does hit/miss statistics for reads only, so the COMA L2 Cache now behaves exactly the same for statistics w.r.t. full/loading/empty lines.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@3978 e97e5017-a994-416e-809d-76780e9f78db