[mgsim] Fixed delegation deadlock bug
As indicated in #280 a circular dependency on the out/in buffers for delegation can cause deadlock. The out-to-in dependency occurs because messages meant for the local core (from the pipeline) are sent to the outgoing buffer after which the routing logic puts it in the input buffer.

The fix is to write local messages directly into the input buffer, bypassing the output buffer and breaking the circular dependency, at the cost of slightly complicating the routing logic.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4015 e97e5017-a994-416e-809d-76780e9f78db