[mgsim] Fixed thread allocation bug.
For group exclusive creates, some processors may not have work to do and mistakingly unreserve a thread whereas they should simply not use the exclusive thread. This caused the deadlock bug in #271 because the erroneous condition was not detected.

Two adjustments have been made: invalid free thread counts are now better detected, and in the situation mentioned above, the thread is no longer unreserved, keeping the thread counts correct.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4020 e97e5017-a994-416e-809d-76780e9f78db