[mgsim] Added per-component clock support.
The simulator now supports components to run at different clock frequencies. This is modeled by having the process, storage and arbitrator list per-clock and only processing clocks if the 
current cycle is their tick time. When creating the clocks, a master simulation frequency is determined such that each clock's cycle time is an integer multiple of the master clock's cycle 
time.

A new configuration block is added with the four frequencies (master, core, memory, DDR) and the 'state' command now accounts for the different clocks.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4024 e97e5017-a994-416e-809d-76780e9f78db