[mgsim] Implemented round-robin thread rescheduler.
The thread scheduler used to pick from the two ready queues based on a single priority-model. If the first list had threads, it would reschedule those, and only then move on to the second one. However, in situations where the active threads can continue for an extended period of time, this causes starvation of the second ready list, which can cause loss of performance due to a lack of work being distributed to other cores (e.g., a remote allocate completes, but the create is never sent because the thread is never rescheduled).

To solve this, the scheduler now uses a round-robin prioritization for the two buffers.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4030 e97e5017-a994-416e-809d-76780e9f78db