[mgsim] Added configuration variable for COMA ring node buffer size.
The variable is called COMA_NodeBufferSize and controls the size of the input and output buffers in the ML COMA's ring networks. This size can have an impact on performance, as a full buffer means that messages on the upper ring cannot shortcut and must travel the long way, via the lower ring(s).

Also, the associativity and number of sets variables have been updated to reflect the actual variable names in config.ini. They have the same value, so no changes in semantics should occur with the default configuration.


git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4104 e97e5017-a994-416e-809d-76780e9f78db