[mgsim] Fixed bug in ML COMA.
Sometimes a request with data can arrive at a root directory after a line has been entirely evicted from the system because all the tokens ended up at root directory. Previously, the root directory assumed that if a read request with data arrived, the line would have to be system. But due to timing issues, this does not have to the be the case.

Therefore, the root directory has been updated to simply 'read' the line from memory again if it's not in the system. Note that it doesn't actually read the line since the request already has the data. It simply allocates a line, creates the tokens and attaches them to the request.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4112 e97e5017-a994-416e-809d-76780e9f78db