[mgsim] Fixed deadlock bug.
The deadlock was caused by a missing create completion writeback.

The completion writeback reached the register file before the create instruction could write back the pending state. Thus, the pending state overwrote the notification writeback.

As the solution, a check has been inserted at the writeback that stalls the notification until the register is Pending.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4195 e97e5017-a994-416e-809d-76780e9f78db