[mgsim-simple-distribution] ISA restructuring.
Some changes have been made to the Alpha and Sparc ISAs. First, the "exact' flag has been moved from the place ID to a new operand of the allocate instructions.

Secondly, the Leon3 ISA merged with Microgrid ISA. Leon3 binaries should now be runnable on the microgrid (not the other way around, though). As part of this merger and to enhance the new places mechanism, new instructions have been added:
* gettid: get the index in the thread table of the current thread.
* getfid: get the index in the family table of the current thread.
* getpid: get the full place ID of the current family.
* getcid: get the global core ID where the current thread is running.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/simple-distribution@4228 e97e5017-a994-416e-809d-76780e9f78db