[mgsim-load-balancing] Implemented load balancing mechanism.
Any allocate with the second least-significant bit (decimal 2) set in its flags operand will now find the least busy core in the target place and allocate the entire family there.

The least busy core is found by looking at the number of used non-exclusive family table entries.

First the allocate goes through the cores in the place, checking if each core's number of used entries is less than the LoadBalanceThreshold configuration option. If so, it allocates there. If all cores in the place are at or above the threshold, it dispatches a remote allocate to the core with the minimum of used entries (which it kept track of as it went through the place).



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/load-balancing@4244 e97e5017-a994-416e-809d-76780e9f78db