[mgsim] Fixed small allocate inconsistency.
A non-exact suspend-allocate behaved as an exact suspend allocate because the suspend flag would be passed to every core, causing them to wait until an entry was available on every core.

Solved by only forwarding the suspend flag to the cores other than the first one if the exact flag is set.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4246 e97e5017-a994-416e-809d-76780e9f78db