[mgsim] Fixed concurrent breeaks.
When multiple threads in a family issue a break, any breaks after the first one assert on the fact that the family has alrady been broken or stopped. Thus, a check has been inserted before breaking a family on a core.

Also, the regression test cases have been updated for break.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4249 e97e5017-a994-416e-809d-76780e9f78db