[mgsim-refactor] Fixed deadlock in COMA.
The deadlock was caused by the top-level ring being full and unable to move because the top-incoming process at a directory could not ever obtain access to the lines, because the bottom-incoming process had priority. In order for the deadlock-freedom in the COMA memory to work, neither process can be given absolute priority. The solution that was to implemented before was a round-robin arbitrator instead of a priority arbitrator. However, its implementation was bugged; it did not properly ignore the "last selected" process when deciding the next one.



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4438 e97e5017-a994-416e-809d-76780e9f78db