[mgsim] Fixed read completion timing bug in DDR.
Read requests in DDR channels were incorrectly regarded as 'done'. Instead of checking that a request's completion time lies 
before the current time, a request was considered done when it's completion time was after the current time.

This causes requests to complete too early, and under high-pressure situations it can cause requests to never complete (when the DDR channel is stalled long enough for the current time to pass the completion time).



git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4525 e97e5017-a994-416e-809d-76780e9f78db