[mgsim] Fix a consistency bug with non-COMA memories.
The non-COMA memories would use a different line size than the cache
line size at the core boundary. Because of this it was possible for a
write request to the middle of a cache line be handled in a different
queue than loads to the start of the cache line. This changeset
prevents this situation by ensuring that the memories use the same
line size as the processor.

git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4672 e97e5017-a994-416e-809d-76780e9f78db