[mgsim] Reinstate the short-latency sampling of real time.
(requested by Irfan.)

The sampling of performance counters 13,14,15 returns wall clock
time. This allows benchmarks and other measurement mechanisms to
sample wall clock time without incurring the expense of a full I/O
access to the RTC device.

This is a feature dedicated to platform analysis, not for realistic
use by programs. It was removed on April 29th, 2011, but is now
reinstated as it is deemed useful by Irfan for simulation performance
analysis.

git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/trunk@4928 e97e5017-a994-416e-809d-76780e9f78db